move user handlers into user package
